server/requirements.txt
pyyaml
python-dotenv
openai
numpy
